requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
transformers==4.35.0
torch>=2.2.0
openai-whisper==20231117
pyannote.audio==2.1.1
numpy==1.26.0
soundfile==0.12.1
openai
streamlit